requirements.txt
fastapi~=0.103.0
requests~=2.31.0
langchain==0.0.275
openai
uvicorn
#pdf 解析类库
pymupdf
pyyaml~=6.0.1
SQLAlchemy~=2.0.20
pymysql